requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.5.0
Django==3.1
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.4.2
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
openapi-codec==1.3.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
zope.interface==5.1.0
